better types and casts

+++ cpc/include/cpc_compressor.hpp
@@ -76,14 +76,14 @@ public:
   size_t low_level_compress_pairs(
       const uint32_t* pair_array, // input
       size_t num_pairs_to_encode,
-      size_t num_base_bits,
+      uint8_t num_base_bits,
       uint32_t* compressed_words // output
   ) const;
 
   void low_level_uncompress_pairs(
       uint32_t* pair_array, // output
       size_t num_pairs_to_decode,
-      size_t num_base_bits,
+      uint8_t num_base_bits,
       const uint32_t* compressed_words, // input
       size_t num_compressed_words // input
   ) const;
@@ -132,12 +132,12 @@ private:
   vector_u32<A> uncompress_surprising_values(const uint32_t* data, size_t data_words, size_t num_pairs, uint8_t lg_k, const A& allocator) const;
   void uncompress_sliding_window(const uint32_t* data, size_t data_words, vector_u8<A>& window, uint8_t lg_k, uint32_t num_coupons) const;
 
-  static size_t safe_length_for_compressed_pair_buf(uint64_t k, size_t num_pairs, size_t num_base_bits);
-  static size_t safe_length_for_compressed_window_buf(uint64_t k);
+  static size_t safe_length_for_compressed_pair_buf(uint32_t k, size_t num_pairs, uint8_t num_base_bits);
+  static size_t safe_length_for_compressed_window_buf(uint32_t k);
   static uint8_t determine_pseudo_phase(uint8_t lg_k, uint64_t c);
 
   static inline vector_u32<A> tricky_get_pairs_from_window(const uint8_t* window, uint32_t k, uint32_t num_pairs_to_get, uint32_t empty_space, const A& allocator);
-  static inline uint64_t golomb_choose_number_of_base_bits(uint64_t k, uint64_t count);
+  static inline uint8_t golomb_choose_number_of_base_bits(uint32_t k, uint64_t count);
 };
 
 } /* namespace datasketches */
